# api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Shovel Heroes Public API
  version: 0.1.0
  description: |
    REST endpoints replacing previous Base44 SDK usage.
    This spec is inferred from the frontend code. Adjust field types and required properties to match your backend.
...
servers:
  - url: https://your.api.server
    description: Production
  - url: http://localhost:8787
    description: Local Dev
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ID:
      type: string
      description: 唯一識別字串 (UUID、ULID 或資料庫主鍵)
      example: "grid_01HZYQ9W123ABCDEF"
    Timestamp:
      type: string
      format: date-time
      description: ISO 8601 時間戳記 (UTC)
      example: "2025-10-02T08:12:30Z"
...
    VolunteerRegistration:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/ID'
        grid_id:
          $ref: '#/components/schemas/ID'
        user_id:
          $ref: '#/components/schemas/ID'
        created_at:
          $ref: '#/components/schemas/Timestamp'
      required: [id, grid_id, user_id]
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: 錯誤訊息
          example: 未授權的請求
        code:
          type: string
          description: 內部錯誤代碼 (可選)
          example: AUTH_REQUIRED
      required: [message]
  parameters:
    PageLimit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
      description: 單頁筆數 (預設 50, 上限 200)
    PageOffset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
      description: 起始位移 (用於分頁)
  responses:
    Unauthorized:
      description: 未授權 (需要 Bearer Token)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 找不到資源
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: 請求參數或內容驗證失敗
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: 伺服器內部錯誤
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /volunteer-registrations:
    get:
      summary: List volunteer registrations
      description: 取得所有志工報名紀錄 (未加分頁僅示範)。
      tags: [VolunteerRegistrations]
      parameters:
        - $ref: '#/components/parameters/PageLimit'
        - $ref: '#/components/parameters/PageOffset'
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/VolunteerRegistration' } } } } }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      summary: Register volunteer
      description: 志工報名參與指定網格。
      tags: [VolunteerRegistrations]
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VolunteerRegistration' }
      responses:
        '201': { description: Created }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
  /volunteer-registrations/{id}:
    delete:
      summary: Cancel volunteer registration
      description: 取消指定志工報名。
      tags: [VolunteerRegistrations]
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/ID' }
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
...